+++ .storybook/vite.config.ts
@@ -2,7 +2,6 @@ import { resolve } from "path";
 import react from "@vitejs/plugin-react";
 import { defineConfig } from "vite";
 
-
 const config = {
   assetsInclude: ["/sb-preview/runtime.js"],
   plugins: [